refactor(parser): extract HTML structure pipeline
